Move create/fork repository from models to modules/repository
